requirements.txt
customtkinter
pymupdf
python-barcode
pyinstaller
